split diff-props into it's own file & add specific test file for it, expand add attributes test
